+++ src/CryptoNoteCore/Blockchain.h
@@ -71,7 +71,7 @@ namespace CryptoNote {
     bool haveTransaction(const Crypto::Hash &id);
     bool haveTransactionKeyImagesAsSpent(const Transaction &tx);
 
-    uint32_t getCurrentBlockchainHeight(); //TODO rename to getCurrentBlockchainSize
+    uint64_t getCurrentBlockchainHeight(); //TODO rename to getCurrentBlockchainSize
     Crypto::Hash getTailId();
     Crypto::Hash getTailId(uint32_t& height);
     difficulty_type getDifficultyForNextBlock();
